Day16: compiler flags and keyboard input

+++ code/handmade.cpp
@@ -11,8 +11,8 @@ static void RenderWeirdGradient(game_offscreen_buffer *Buffer, int BlueOffset, i
             // Spreads grandient across whole window
             // uint8 Green = (( ((double)X) / Buffer.Width) * 256) + XOffset;
             // uint8 Blue = (( ((double)Y) / Buffer.Height) * 256) + YOffset;
-            uint8 Green = X + BlueOffset;
-            uint8 Blue = Y + GreenOffset;
+            uint8 Green = (uint8)(X + BlueOffset);
+            uint8 Blue = (uint8)(Y + GreenOffset);
 
             *Pixel++ = ((Green << 8) | Blue);
         }
@@ -72,9 +72,9 @@ static void GameUpdateAndRender(game_memory *Memory,
     game_controller_input *Input0 = &Input->Controllers[0];
     if (Input0->IsAnalog)
     {
-        GameState->BlueOffset += (int)4.0f * (Input0->EndX);
-        GameState->GreenOffset += (int)4.0f * (Input0->EndY);
-        GameState->ToneHz = 256 + (int)128.0f * (Input0->EndY);
+        GameState->BlueOffset += (int)(4.0f * (Input0->EndX));
+        GameState->GreenOffset += (int)(4.0f * (Input0->EndY));
+        GameState->ToneHz = 256 + (int)(128.0f * (Input0->EndY));
 
     }
     else
